+++ lshell/checkconfig.py
@@ -570,6 +570,7 @@ class CheckConfig:
             "quiet",
             "security_audit_json",
             "max_sessions_per_user",
+            "max_background_jobs",
         ]:
             try:
                 if len(self.conf_raw[item]) == 0: