+++ lshell/configschema.py
@@ -43,6 +43,7 @@ INT_VALUE_KEYS = {
     "loglevel",
     "security_audit_json",
     "max_sessions_per_user",
+    "max_background_jobs",
 }
 DICT_VALUE_KEYS = {"aliases", "env_vars", "messages"}
 STRING_VALUE_KEYS = {